refactor: rename some files. change name of action and reducer directory. change formulas to historylist. make destructuring actions
