+++ src/components/Calculator/CalculatorFC.jsx
@@ -5,7 +5,7 @@ import {
     drawDisplay,
     drawHistory,
     drawHistoryDisplay,
-} from '@actions/displayActions';
+} from '@actions/display';
 import AddCommand from '@command/AddCommand';
 import calc from '@command/command';
 import DivideCommand from '@command/DivideCommand';
