+++ src/components/Header/Switcher/SwitcherFC.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { switchTheme } from '@actions/controlsActions';
+import switchTheme from '@actions/control';
 
 import { Handler, WrapperSwitcher } from './styled';
