+++ src/store/actions/control.jsx
@@ -1,0 +1,8 @@
+import { SWITCH_THEME } from './type';
+
+const switchTheme = (payload) => ({
+    type: SWITCH_THEME,
+    payload,
+});
+
+export default switchTheme;
