--- src/store/actions/controlsActions.jsx
@@ -1,10 +1,0 @@
-import { MAKE_ACTUAL, SWITCH_THEME } from './type';
-
-export const setActualDig = (payload) => ({
-    type: MAKE_ACTUAL,
-    payload,
-});
-export const switchTheme = (payload) => ({
-    type: SWITCH_THEME,
-    payload,
-});
