+++ src/store/actions/type.js
@@ -1,9 +1,8 @@
-export const MAKE_ACTUAL = 'CONTROL/MAKE_ACTUAL';
-export const SWITCH_THEME = 'CONTROL/SWITCH_THEME';
+export const SWITCH_THEME = 'SWITCH_THEME';
 
-export const DRAW_DISPLAY = 'DISPLAY/DRAW_DISPLAY';
-export const DRAW_HISTORY_DISPLAY = 'DISPLAY/DRAW_HISTORY_DISPLAY';
-export const DRAW_HISTORY = 'DISPLAY/DRAW_HISTORY';
-export const CLEAR_DISPLAY = 'DISPLAY/CLEAR_DISPLAY';
-export const CLEAR_HISTORY = 'DISPLAY/CLEAR_HISTORY';
-export const CLEAR_ALL = 'DISPLAY/CLEAR_ALL';
+export const DRAW_DISPLAY = 'DRAW_DISPLAY';
+export const DRAW_HISTORY_DISPLAY = 'DRAW_HISTORY_DISPLAY';
+export const DRAW_HISTORY = 'DRAW_HISTORY';
+export const CLEAR_DISPLAY = 'CLEAR_DISPLAY';
+export const CLEAR_HISTORY = 'CLEAR_HISTORY';
+export const CLEAR_ALL = 'CLEAR_ALL';
