+++ src/store/store.jsx
@@ -2,8 +2,8 @@ import { combineReducers, legacy_createStore as createStore } from 'redux';
 import { persistReducer, persistStore } from 'redux-persist';
 import storage from 'redux-persist/lib/storage';
 
-import controlReducer from './reducers/controlReducer';
-import displayReducer from './reducers/displayReducer';
+import controlReducer from './reducers/control';
+import displayReducer from './reducers/display';
 
 const persistConfig = {
     key: 'root',
